moves cubedata request to axios, only opens on dimensioneditor
